server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.121.2",
    "motor>=3.7.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "bcrypt==4.3.0",
    "pydantic-settings>=2.0.0",
    "passlib>=1.7.4",
    "google-generativeai>=0.3.0",
]
